requirements-dev.txt
pytest==6.2.5
Pillow==10.1.0
requests==2.25.1
boto3==1.28.75
numpy==1.26.2